Middlewares added
errorHandler, requestLogger and unknownEndpoint added
The application is working perfectly with the improvement.

errorHandler - Handles the errors related to Cast Error or Validation
error and forward any other error to the unknownEndpoint.

requestLogger - Log all the request information to the console.

unknownEndpoint - Handles the errors passed by the next function sent
from the errorHandler and responde with the error - unknownEndpoint